requirements.txt
pyaml
python-dotenv
flask
flask_cors
requests